Keyring

Documentation for Keyring.

Keyring.Credential

Keyring.Windows.WindowsCredentialManager

Keyring.DEFAULT_CREDENTIAL_STORE

Keyring.get_credential

Keyring.set_credential

Keyring.Credential — Type

Credential{U,T,V}(target,username,secret) <: AbstractCredential

A credential object that stores the target, username, and secret. Access the fields using the dot syntax, e.g. c.target, c.username, c.secret.

Keyring.DEFAULT_CREDENTIAL_STORE — Method

DEFAULT_CREDENTIAL_STORE()::T<:AbstractCredentialStore

Returns the default credential store for the current system.

Windows: Windows.WindowsCredentialManager

macOS: TBD

Linux: TBD

Keyring.get_credential — Method

get_credential(target::AbstractString)
get_credential(store::AbstractCredentialStore, target::AbstractString)

Returns either a Credential or Some(nothing) if the credential is not found.

If not given, store defaults to the value of DEFAULT_CREDENTIAL_STORE.

Examples

julia> set_credential("my_target", "my_username", "my_secret")
julia> c = get_credential("my_target")

Keyring.set_credential — Method

set_credential(target::AbstractString, username::AbstractString, secret::AbstractString)
set_credential(store::AbstractCredentialStore, target::AbstractString, username::AbstractString, secret::AbstractString)

Sets the value and returns nothing.

If not given, store defaults to the value of DEFAULT_CREDENTIAL_STORE.

Keyring.Windows.WindowsCredentialManager — Type

WindowsCredentialManager <: AbstractCredentialStore

This interacts with the Windows Credential Manager via the advapi32.dll (Advanced Windows 32 Base API) library available on Windows systems.